Python-old/requirements.txt
discord.py==2.5.2
pymongo==4.11.2
python-dotenv==1.0.1
pytz==2025.1
Requests==2.32.3
schedule==1.2.2
flask==3.1.0